# Conflicts:
#	qgcresources.qrc
#	src/FlightDisplay/FlightDisplayView.qml
#	src/MissionManager/MissionController.cc
#	src/MissionManager/SurveyMissionItem.cc
#	src/MissionManager/SurveyMissionItem.h
#	src/PlanView/PlanView.qml
#	src/QGCApplication.cc
#	src/Settings/AppSettings.h
#	src/Vehicle/Vehicle.h

Note that SurveyMissionItem (.h/.cc) was changed in such a way that all previous changes were lost. Dynamic computation of its bounding box is missing and needs to be added into whatever is now handling these.

+++ src/QGC.h
@@ -33,7 +33,7 @@ quint64 groundTimeMilliseconds();
  */
 qreal groundTimeSeconds();
 /** @brief Returns the angle limited to -pi - pi */
-float limitAngleToPMPIf(float angle);
+float limitAngleToPMPIf(double angle);
 /** @brief Returns the angle limited to -pi - pi */
 double limitAngleToPMPId(double angle);
